Fixed sema and shmem library fail flag behaviors
In earlier code, `remove` and `close` functions in `ipc/sema` and
`ipc/shmem` libraries ignored the `fail` flag as no action is to be
taken even when errors occur on calling these functions. Fixed this, so
that these functions do set the `fail` flag, as it is still useful for
logging purposes.